requirements.txt
# ── Django Core ──
django==4.2.7
django-crispy-forms==2.1
crispy-bootstrap5==0.7
gunicorn==22.0.0

# ── AI / Agentic System ──
langchain>=0.3.20
langchain-community>=0.3.20
langchain-groq>=0.2.0
groq>=0.4.0
faiss-cpu==1.7.4
sentence-transformers>=2.2.0

# ── Payments ──
stripe==7.0.0

# ── Utilities ──
python-dotenv==1.0.0
Pillow==11.3.0

# ── Background Tasks ──
celery>=5.3.6
redis>=5.0.1

# ── Data / Visualization (optional, for diagram generation) ──
matplotlib==3.7.2
seaborn==0.12.2
numpy<2.0.0,>=1.22.4
django-ratelimit>=4.1.0
django-csp>=4.0
sentry-sdk>=1.40.6

langchain_huggingface